+++ src/commands/migrate.ts
@@ -11,7 +11,7 @@ export default async function TheMigrator(project: string, target: MANAGERS) {
             await LogStuff("Please wait (this will take a while)...", "working");
             try {
                 await LogStuff("Removing node_modules (1/3)...", "working");
-                await Deno.remove(JoinPaths(target, "node_modules"), {
+                await Deno.remove(await JoinPaths(target, "node_modules"), {
                     recursive: true,
                 });
             } catch (e) {
@@ -28,7 +28,7 @@ export default async function TheMigrator(project: string, target: MANAGERS) {
 
             try {
                 await LogStuff("Removing previous lockfile (3/3)...", "working");
-                await Deno.remove(JoinPaths(target, remove));
+                await Deno.remove(await JoinPaths(target, remove));
             } catch (e) {
                 throw e;
             }
@@ -38,20 +38,20 @@ export default async function TheMigrator(project: string, target: MANAGERS) {
     }
 
     try {
-        const workingProject = ParsePath("path", project) as string;
+        const workingProject = await ParsePath("path", project) as string;
         const workingTarget = target.toLowerCase().trimEnd().trimStart();
 
         if (!(["pnpm", "npm", "yarn"].includes(workingTarget))) {
             throw new Error(`${workingTarget} is not a valid target. Use either "pnpm", "npm", or "yarn".`);
         }
 
-        const isNpm = await CheckForPath(JoinPaths(workingProject, "package-lock.json"));
-        const isPnpm = await CheckForPath(JoinPaths(workingProject, "pnpm-lock.yaml"));
-        const isYarn = await CheckForPath(JoinPaths(workingProject, "yarn.lock"));
+        const isNpm = await CheckForPath(await JoinPaths(workingProject, "package-lock.json"));
+        const isPnpm = await CheckForPath(await JoinPaths(workingProject, "pnpm-lock.yaml"));
+        const isYarn = await CheckForPath(await JoinPaths(workingProject, "yarn.lock"));
 
         if (!isNpm && !isPnpm && !isYarn) throw new Error("We weren't able to find a valid lockfile over here.");
 
-        if (!(await CheckForPath(JoinPaths(target, "package.json")))) {
+        if (!(await CheckForPath(await JoinPaths(target, "package.json")))) {
             throw new Error("No package.json found, cannot migrate. How will we install your modules without it?");
         }