+++ src/commands/stats.ts
@@ -9,9 +9,9 @@ export default async function TheStatistics() {
     await LogStuff(`This is ${I_LIKE_JS.MFLY} going to take a while. Have a coffee meanwhile!`, "bruh");
 
     for (const project of projects) {
-        const workingPath = ParsePath("path", project) as string;
+        const workingPath = await ParsePath("path", project) as string;
 
-        const nodeGarbagePath = JoinPaths(workingPath, "node_modules");
+        const nodeGarbagePath = await JoinPaths(workingPath, "node_modules");
 
         if (!(await CheckForPath(workingPath))) {
             await LogStuff(`${workingPath} doesn't exist?`, "warn");